geocat-ogc-services-checker/requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.3
geopycat @ git+https://github.com/geoadmin/lib-geopycat.git
